% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointGHS.r
\name{boot_and_joint_parallel_iteration}
\alias{boot_and_joint_parallel_iteration}
\title{Perform bootstrap iterarion}
\usage{
boot_and_joint_parallel_iteration(
  X,
  S,
  n.vals,
  p,
  K,
  theta,
  sigma,
  Lambda_sq,
  tau_sq,
  tau_sq_min,
  tau_sq_stepsize,
  tau_sq_max,
  epsilon,
  maxitr,
  AIC_eps,
  perform_boot,
  stop_overflow
)
}
\arguments{
\item{X}{list of the \eqn{K} observed \eqn{n_k} by \eqn{p} data matrices}

\item{S}{list of the \eqn{p} by \eqn{p} scatter matrices of the \eqn{K} data sets}

\item{n.vals}{vector of length \eqn{K} of the number of obervations in each data set.}

\item{p}{number of nodes}

\item{K}{number of networks}

\item{theta}{list of the \eqn{K} initial \eqn{p} by \eqn{p} precision matrices. Optional argument}

\item{sigma}{list of the \eqn{K} initial \eqn{p} by \eqn{p} covariance matrices. Optional argument}

\item{Lambda_sq}{list of the \eqn{K} initial \eqn{p} by \eqn{p} matrices of squared local shrinkage parameters. Optional argument}

\item{tau_sq}{initial values of squared global shrinkage parameters. A vector of length \eqn{K}, or a single value to be used for all networks.}

\item{tau_sq_min}{if \code{AIC_selection=TRUE}, the smallest value of \code{tau_sq} to consider}

\item{tau_sq_stepsize}{if \code{AIC_selection=TRUE}, the step-size to use in the grid for \code{tau_sq}. Optional argument}

\item{tau_sq_max}{if \code{AIC_selection=TRUE}, the largest value of \code{tau_sq} to consider}

\item{epsilon}{tolerance for the convergence assessment}

\item{maxitr}{maximum number of iterations}

\item{AIC_eps}{if \code{AIC_selection=TRUE}, the tolerance for the AIC convergence assessment}

\item{perform_boot}{should bootstrapping be performed? if 1, yes, else perform joint}

\item{stop_overflow}{should measures be taken to avoid overflow? Only necessary for very dense graphs with many large nonzero elements.}
}
\description{
This function performs one Bayesian bootstrap iteration
}
\keyword{internal}
