% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_functions.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Find the confusion matrix of a graph and an estimate of it}
\usage{
confusion.matrix(g, g.hat)
}
\arguments{
\item{g}{The adjacency matrix of the true graph.}

\item{g.hat}{The adjacency matrix of the estimated graph.}
}
\value{
The \eqn{2} by \eqn{2} confusion matrix. Element \code{[1,1]} is the number of true positives, element \code{[1,2]} the number of false positives, element  \code{[2,1]} the number of false negatives and  \code{[2,2]} the number of true negatives.
}
\description{
Finds the confusion matrix of an inferred graph and the graph it aims to estimate.
}
\examples{

# example 1: simple example with two unrelated matrices
# generate data and prior matrix
set.seed(123)
x1 <- matrix(rnorm(40 * 20), ncol = 20)
x2 <- matrix(rnorm(40 * 20), ncol = 20)
# the precision matrices
prec.mat.1 <- solve(var(x1))
prec.mat.2 <- solve(var(x2))
# make them sparse by threshold values. Do not make diagonal elements zero.
prec.mat.1[which(abs(prec.mat.1) < 1.5 & !diag(ncol(prec.mat.1)))] <- 0
prec.mat.2[which(abs(prec.mat.2) < 1.5 & !diag(ncol(prec.mat.2)))] <- 0
confusion.matrix(prec.mat.1 != 0, prec.mat.2 != 0)
}
\seealso{
\code{\link{jointGHS}}
}
\author{
Camilla Lingjaerde
}
