% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGHS.r
\name{fastGHS}
\alias{fastGHS}
\title{Perform GHS with ECM/ICM}
\usage{
fastGHS(
  X,
  theta = NULL,
  sigma = NULL,
  Lambda_sq = NULL,
  tau_sq = NULL,
  method = "ECM",
  epsilon = 1e-05,
  maxitr = 1e+05,
  verbose = TRUE,
  savepath = FALSE,
  group = NULL,
  save_Q = F,
  fix_tau = FALSE,
  stop_underflow = FALSE
)
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} matrix of data}

\item{theta}{initial value of the \eqn{p} by \eqn{p} precision matrix}

\item{sigma}{initial value of the \eqn{p} by \eqn{p} covariance matrix}

\item{Lambda_sq}{initial value of matrix of squared local shrinkage parameters}

\item{tau_sq}{initial value of squared global shrinkage parameter. If variables are grouped, a \eqn{ngroup} by \eqn{ngroup} matrix}

\item{method}{the method to use. Default is \eqn{ECM}. Other options include \eqn{ICM}}

\item{epsilon}{tolerance for the convergence assessment}

\item{maxitr}{maximum number of iterations}

\item{verbose}{logical indicator of printing information at each iteration}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for p<200}

\item{group}{grouping information, if variables are grouped. If provided, a vector of length \eqn{p} giving the group of each variable as a string, character or number.}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{fix_tau}{logical. Should tau be fixed?}

\item{stop_underflow}{should underflow be avoided by never allowing doubles to be smaller than the machine precision?}
}
\value{
a fitted EMGS object
}
\description{
This function performs expectation-conditional-maximation or iterated conditional mode estimation for the graphical horseshoe
}
