% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_functions.R
\name{gaussianAIC}
\alias{gaussianAIC}
\title{Calculates the AIC score of an estimated precision matrix}
\usage{
gaussianAIC(sample.cov, theta, n)
}
\arguments{
\item{sample.cov}{The sample covariance matrix of the observed data.}

\item{theta}{The estimated precision matrix.}

\item{n}{The number of observations.}
}
\value{
The AIC score.
}
\description{
Finds the AIC score of an estimated precision matrix given a sample covariance matrix, in the setting of Gaussian graphical models.
}
\examples{

# example 1: simple example where we check the AIC score of the true precision matrix
# generate data
set.seed(123)
n <- 80
p <- 100
dat <- huge::huge.generator(n = n, d = p, graph = "scale-free")
prec.mat <- dat$omega # true precision matrix
gaussianAIC(var(dat$data), prec.mat, n)
}
\seealso{
\code{\link{jointGHS}}
}
\author{
Camilla Lingjaerde
}
