% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_functions.R
\name{gaussianloglik}
\alias{gaussianloglik}
\title{Calculates the log likelihood of a precision matrix}
\usage{
gaussianloglik(sample.cov, theta, n)
}
\arguments{
\item{sample.cov}{The sample covariance matrix of the observed data.}

\item{theta}{The estimated precision matrix.}

\item{n}{The number of observations.}
}
\value{
Numerical value of the multivariate Gaussian profile log likelihood of the precision matrix.
}
\description{
Finds the the profile log likelihood of a precision matrix given a sample covariance matrix, in the setting of Gaussian graphical models.
}
\examples{

# example 1: simple example where we check the log likelihood of the true precision matrix
# generate data
set.seed(123)
n <- 80
p <- 100
dat <- huge::huge.generator(n = n, d = p, graph = "scale-free")
prec.mat <- dat$omega # true precision matrix
gaussianloglik(var(dat$data), prec.mat, n)
}
\seealso{
\code{\link{jointGHS}}
}
\author{
Camilla Lingjaerde
}
\keyword{internal}
