% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointGHS.r
\name{jointGHS}
\alias{jointGHS}
\title{Perform joint GHS with ECM/ICM}
\usage{
jointGHS(
  X,
  theta = NULL,
  sigma = NULL,
  Lambda_sq = NULL,
  tau_sq = NULL,
  method = "ECM",
  epsilon = 1e-05,
  maxitr = 1e+05,
  scale = FALSE,
  verbose = TRUE,
  savepath = FALSE,
  save_Q = FALSE,
  fix_tau = TRUE
)
}
\arguments{
\item{X}{list of the \eqn{K} observed \eqn{n_k} by \eqn{p} data matrices}

\item{theta}{list of the \eqn{K} initial \eqn{p} by \eqn{p} precision matrices. Optional argument}

\item{sigma}{list of the \eqn{K} initial \eqn{p} by \eqn{p} covariance matrices. Optional argument}

\item{Lambda_sq}{list of the \eqn{K} initial \eqn{p} by \eqn{p} matrices of squared local shrinkage parameters. Optional argument}

\item{tau_sq}{initial values of squared global shrinkage parameters. A vector of length \eqn{K}, or a single value to be used for all networks}

\item{method}{the method to use. Default is \eqn{ECM}. Other options include \eqn{ICM}}

\item{epsilon}{tolerance for the convergence assessment}

\item{maxitr}{maximum number of iterations}

\item{scale}{should variables be scaled?}

\item{verbose}{logical indicator of printing information at each iteration}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for p<200}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{fix_tau}{logical. Should tau be fixed?}
}
\value{
a fitted EMGS object
}
\description{
This function performs expectation-conditional-maximation or iterated conditional mode estimation for the joint graphical horseshoe
}
