% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointGHS.r
\name{plot.jointGHS}
\alias{plot.jointGHS}
\title{Plot function for S3 class "jointGHS"}
\usage{
\method{plot}{jointGHS}(x, k = 1, plot_boot = TRUE, edges = NA, quantiles = c(0.025, 0.975))
}
\arguments{
\item{x}{An object with S3 class \code{"jointGHS"}}

\item{k}{the network to plot}

\item{plot_boot}{should the Bayesian bootstrap sample posterior of some of the Lambda_sq be shown? If \code{FALSE}, the jointGHS networks are visualized instead}

\item{edges}{the edges for which to show the Bayesian bootstrap posterior distribution. An matrix with \eqn{2} columns. If not specified, \eqn{15} edges are selected at random amongst the infererred jointGHS edge set}

\item{quantiles}{the quantiles to show when plotting the Bayesian bootstrap posterior distribution of the local scale parameters}
}
\description{
This function provides functionality for plotting the results of the jointGHS. Can be used to evaluate the local scales \code{Lambda_sq} of a network from the joint analysis against their Bayesian bootstrap distributions, and to visualize the inferred networks
}
\seealso{
\code{\link{jointGHS}}, \code{\link{print.jointGHS}}
}
