% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointGHS.r
\name{plot_jointGHS}
\alias{plot_jointGHS}
\title{Plot function for S3 class "jointGHS"}
\usage{
plot_jointGHS(
  x,
  k = 1,
  plot_boot = TRUE,
  edges = NA,
  percentile = 95,
  quantiles = NULL,
  show_single = FALSE,
  true_theta = NULL
)
}
\arguments{
\item{x}{An object with S3 class \code{"jointGHS"}}

\item{k}{the network to plot}

\item{plot_boot}{should the Bayesian bootstrap sample posterior of some of the Lambda_sq be shown? If \code{FALSE}, the jointGHS networks are visualized instead}

\item{edges}{the edges for which to show the Bayesian bootstrap posterior distribution. An matrix with \eqn{2} columns. If not specified, \eqn{15} edges are selected at random amongst the infererred jointGHS edge set}

\item{percentile}{the percentile to show when plotting the Bayesian bootstrap posterior distribution of the local scale parameters.}

\item{quantiles}{if provided, the quantiles to show when plotting the Bayesian bootstrap posterior distribution of the local scale parameters. Overrides \code{percentile} argument}

\item{show_single}{should the estimate from the single network analysis be included as a bar?}

\item{true_theta}{if provided, the 'true' value of \code{theta} to include as a bar in the plot}
}
\description{
This function provides functionality for plotting the results of the jointGHS. Can be used to evaluate the local scales \code{Lambda_sq} of a network from the joint analysis against their Bayesian bootstrap distributions, and to visualize the inferred networks
}
\seealso{
\code{\link{jointGHS}}, \code{\link{print_jointGHS}}
}
