% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_functions.R
\name{precision}
\alias{precision}
\title{Find the precision of an estimated graph}
\usage{
precision(g, g.hat)
}
\arguments{
\item{g}{The adjacency matrix of the true graph.}

\item{g.hat}{The adjacency matrix of the estimated graph.}
}
\value{
Numeric precision.
}
\description{
Computes the precision of an inferred graph in terms of predicting the edges of the true one. The measure has a value in \eqn{[0,1]} where a value of \eqn{1} means perfect prediction while a value of \eqn{0} means completely wrong.
}
\examples{

# example 1: simple example with two unrelated matrices
# generate data and prior matrix
set.seed(123)
x1 <- matrix(rnorm(40 * 20), ncol = 20)
x2 <- matrix(rnorm(40 * 20), ncol = 20)
# the precision matrices
prec.mat.1 <- solve(var(x1))
prec.mat.2 <- solve(var(x2))
# make them sparse by threshold values. Do not make diagonal elements zero.
prec.mat.1[which(abs(prec.mat.1) < 1.5 & !diag(ncol(prec.mat.1)))] <- 0
prec.mat.2[which(abs(prec.mat.2) < 1.5 & !diag(ncol(prec.mat.2)))] <- 0
precision(prec.mat.1 != 0, prec.mat.2 != 0)
}
\seealso{
\code{\link{jointGHS}} and \code{\link{confusion.matrix}}
}
\author{
Camilla Lingjaerde
}
