% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointGHS.r
\name{print_jointGHS}
\alias{print_jointGHS}
\title{Print function for S3 class "jointGHS"}
\usage{
print_jointGHS(x, k = 1, edges = NA, percentile = 95, return_df = F)
}
\arguments{
\item{x}{an object with S3 class \code{"jointGHS"}}

\item{k}{the network to print the results for}

\item{edges}{the edges for which to show the Bayesian bootstrap posterior distribution. A matrix with \eqn{2} columns, or alternatively the string \code{'all'}. If not provided, \eqn{16} edges are selected at random amongst the infererred jointGHS edge set}

\item{percentile}{the empirical percentile of the Bayesian bootstrap posterior distribution of the local scale parameters we want to show. Can also be a vector of length \eqn{2} for two-sided intervals}

\item{return_df}{should the results be returned as a data frame? Default \code{FALSE}}
}
\value{
object of class \code{"data.frame"}
}
\description{
This function provides functionality for printing a summary of the results of the jointGHS Bayesian bootstrap check. The local scale parameters from the joint analysis are checked against their Bayesian bootstrap distribution in the single network version
}
\seealso{
\code{\link{jointGHS}}, \code{\link{plot_jointGHS}}
}
