% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_functions.R
\name{sparsity}
\alias{sparsity}
\title{The sparsity of a graph}
\usage{
sparsity(g)
}
\arguments{
\item{g}{The adjacency matrix of the graph.}
}
\value{
Numeric sparsity.
}
\description{
Finds the sparsity of a graph. The measure has a value in \eqn{[0,1]} where a value of \eqn{1} means there are edges between all nodes while a value of \eqn{0} means there are no edges in the graph.
}
\examples{

# example 1: simple example with two unrelated matrices
# generate data
set.seed(123)
n <- 80
p <- 100
dat <- huge::huge.generator(n = n, d = p, graph = "scale-free")
sparsity(dat$theta)
}
\seealso{
\code{\link{jointGHS}} and \code{\link{confusion.matrix}}
}
\author{
Camilla Lingjaerde
}
