% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ECM_GHS}
\alias{ECM_GHS}
\title{Perform jointGHS with ECM}
\usage{
ECM_GHS(
  S,
  theta,
  sigma,
  Lambda_sq,
  N,
  M,
  K,
  epsilon,
  verbose,
  maxitr,
  savepath,
  save_Q,
  tau_sq,
  use_ICM = FALSE,
  fix_tau = TRUE
)
}
\arguments{
\item{S}{cube of the \eqn{p} by \eqn{p} scatter matrices of the \eqn{K} data sets}

\item{theta}{cube of the \eqn{K} initial \eqn{p} by \eqn{p} precision matrices}

\item{sigma}{cube of the \eqn{K} initial \eqn{p} by \eqn{p} covariance matrices}

\item{Lambda_sq}{cube of the \eqn{K} initial \eqn{p} by \eqn{p} matrices of squared local shrinkage parameters}

\item{N}{vector of length \eqn{K} of the number of obervations in each data set.}

\item{M}{the number of variables/nodes. Integer}

\item{K}{the number of networks. Integer}

\item{epsilon}{tolerance for the convergence assessment}

\item{verbose}{logical indicator of printing information at each iteration}

\item{maxitr}{maximum number of iterations}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for p<200}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{tau_sq}{vector of initial values of the \eqn{K} squared global shrinkage parameters}

\item{use_ICM}{logical. Should ICM be used instead of ECM? Default value is false}

\item{fix_tau}{logical. Should tau be fixed?}
}
\value{
A List with resulting ECM estimates, and saved path and objective function convergence information if requested
}
\description{
This function performs expectation-conditional-maximation or iterated conditional mode estimation for the joint graphical horseshoe
}
