% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointGHS.r
\name{jointGHS}
\alias{jointGHS}
\title{Perform joint GHS with ECM/ICM}
\usage{
jointGHS(
  X,
  theta = NULL,
  sigma = NULL,
  Lambda_sq = NULL,
  tau_sq = NULL,
  method = "ECM",
  AIC_selection = TRUE,
  AIC_eps = 0.1,
  tau_sq_min = 0.001,
  tau_sq_stepsize = NULL,
  tau_sq_max = 20,
  epsilon = 1e-05,
  maxitr = 1000,
  scale = TRUE,
  verbose = TRUE,
  savepath = FALSE,
  save_Q = FALSE,
  fix_tau = FALSE,
  boot_check = FALSE,
  B = 100,
  nCores = 5,
  boot_lambda = TRUE
)
}
\arguments{
\item{X}{list of the \eqn{K} observed \eqn{n_k} by \eqn{p} data matrices}

\item{theta}{list of the \eqn{K} initial \eqn{p} by \eqn{p} precision matrices. Optional argument}

\item{sigma}{list of the \eqn{K} initial \eqn{p} by \eqn{p} covariance matrices. Optional argument}

\item{Lambda_sq}{list of the \eqn{K} initial \eqn{p} by \eqn{p} matrices of squared local shrinkage parameters. Optional argument}

\item{tau_sq}{initial values of squared global shrinkage parameters. A vector of length \eqn{K}, or a single value to be used for all networks.}

\item{method}{the method to use. Default is \code{ECM}. Other options include \code{ICM}}

\item{AIC_selection}{logical. Should the global shrinkage parameters be selected with AIC? Default \code{TRUE}}

\item{AIC_eps}{if \code{AIC_selection=TRUE}, the tolerance for the AIC convergence assessment}

\item{tau_sq_min}{if \code{AIC_selection=TRUE}, the smallest value of \code{tau_sq} to consider}

\item{tau_sq_stepsize}{if \code{AIC_selection=TRUE}, the step-size to use in the grid for \code{tau_sq}. Optional argument}

\item{tau_sq_max}{if \code{AIC_selection=TRUE}, the largest value of \code{tau_sq} to consider}

\item{epsilon}{tolerance for the convergence assessment}

\item{maxitr}{maximum number of iterations}

\item{scale}{should variables be scaled? Default \code{TRUE}}

\item{verbose}{logical indicator of printing information at each iteration}

\item{savepath}{logical indicator of saving the estimator at each iteration in the ECM algorithm. Only available for \eqn{p<200}}

\item{save_Q}{should the value of the objective function at each step be saved?}

\item{fix_tau}{logical. Should \code{tau_sq} be fixed? Default \code{FALSE}}

\item{boot_check}{should bootstrapping be performed to check the joint results?}

\item{B}{if \code{boot_check=TRUE}, the number of bootstrap samples to draw}

\item{nCores}{if \code{boot_check=TRUE}, how many cores should be used for the bootstrap sampling?}

\item{boot_lambda}{should \code{Lambda_sq} be bootstrapped? If \code{FALSE}, \code{theta} is bootstrapped instead}
}
\value{
a fitted \code{jointGHS} object
}
\description{
This function performs expectation-conditional-maximation or iterated conditional mode estimation for the joint graphical horseshoe
}
\seealso{
\code{\link{plot.jointGHS}}, \code{\link{print.jointGHS}}
}
